Filtering and sorting (#4)
Add filtering by grade, and sorting by rating.

This made handling paste events and climb generation responses slightly
awkward or impossible to navigate when there were multiple climbs.

To work around this, there's an override for the filters so that the
contents of the latest paste/generation are always shown instead of the
filtered routes. An indicator appears near the grade filter which you
can tap to get back to the filtered climbs.